+++ src/main/java/com/martinsil/pvzcraft/entity/client/RegularZombieRenderer.java
@@ -6,6 +6,7 @@ import net.minecraft.client.render.VertexConsumerProvider;
 import net.minecraft.client.render.entity.EntityRendererFactory;
 import net.minecraft.client.render.entity.MobEntityRenderer;
 import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.util.Identifier;
 
 public class RegularZombieRenderer extends MobEntityRenderer<RegularZombieEntity, RegularZombieModel<RegularZombieEntity>> {