+++ src/main/java/com/martinsil/pvzcraft/map/MapManager.java
@@ -1,0 +1,14 @@
+package com.martinsil.pvzcraft.map;
+
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.server.world.ServerWorld;
+
+public class MapManager {
+
+    public static void clearWorld(ServerWorld world) {
+        for (Entity entity : world.iterateEntities())
+            if (!(entity instanceof PlayerEntity))
+                entity.discard();
+    }
+}